+++ R/with_progress.R
@@ -36,6 +36,9 @@
 #' @example incl/with_progress.R
 #'
 #' @details
+#' If the global progression handler is enabled, it is temporarily disabled
+#' while evaluating the `expr` expression.
+#'
 #' **IMPORTANT: This function is meant for end users only.  It should not
 #' be used by R packages, which only task is to _signal_ progress updates,
 #' not to decide if, when, and how progress should be reported.**
@@ -76,7 +79,11 @@ with_progress <- function(expr, handlers = progressr::handlers(), cleanup = TRUE
     message("with_progress() ...")
     on.exit(message("with_progress() ... done"), add = TRUE)
   }
-  
+
+  ## Deactive global progression handler while using with_progress()
+  global_progression_handler(FALSE)
+  on.exit(global_progression_handler(TRUE), add = TRUE)
+
   ## FIXME: With zero handlers, progression conditions will be
   ##        passed on upstream just as without with_progress().
   ##        Is that what we want? /HB 2019-05-17